All endpoints now connected and functional, needs some cleanup and test fixes
